+++ src/main/kotlin/dev/digamma/shard/util/Plugin.kt
@@ -1,0 +1,5 @@
+package dev.digamma.shard.util
+
+import com.intellij.openapi.extensions.PluginId
+
+val PLUGIN_ID = PluginId.getId("dev.digamma.shard")